Refactored api file structure to seperate routers
