+++ api/src/routers/user.py
@@ -1,0 +1,3 @@
+from fastapi import APIRouter
+
+router = APIRouter()
